fix: fixed route for auth-callback. was under login. just swapped tsx pages
